Added index.html and stylesheet.css
Captain America shield and animation made by html and css. I am practicing web development and design.

A filter was added on the shield to give it more depth - darker on the 2 side of the shield that is fixed while the shield rotates. Feels more accurate how light interacts with objects if the source doesn't move with the object shined upon.